refactor(rip): move docLabel to the backend, where coverage is measured
packages/shared is not one of the five coverage-gated workspaces: no test
script, no runner, zero tests. A branching function there is unmeasurable
and cannot meet the repo's 80% per-file branch floor, which a passing test
run had concealed — the test ran fine from the backend against dist.

An existing project note already stated the invariant this broke: shared
is "types plus two seed/data modules with no functions and no branches".
docLabel was the first function to violate it.

Shared keeps the type declarations, which have no branches. doc-label.ts
measures at 100% branch coverage in the backend.

Better placement regardless: docLabel has no frontend consumer. The parser
resolves labels server-side and ships the model with `doc` already filled
in, so the frontend never calls it.
